added tests for seraching users
